src/app/trips-api/requirements.txt
flask
pymongo
kafka-python
python-dotenv
requests
numpy
bson